Separate production and test code
