+++ nested-lambdas/javaspec-client/src/main/java/info/javaspec/client/GameState.java
@@ -1,0 +1,10 @@
+package info.javaspec.client;
+
+import java.util.List;
+
+public interface GameState {
+	boolean hasWon(String player);
+	boolean isOver();
+	List<String> nextMoves();
+	GameState nextState(String move);
+}
