+++ nested-lambdas/javaspec-client/src/test/java/info/javaspec/client/MinimaxSpecs.java
@@ -81,51 +81,7 @@ public class MinimaxSpecs implements SpecClass {
 		return game;
 	}
 
-	public static class Minimax {
-		private final String maximizer;
-		private final String minimizer;
-
-		public Minimax(String maximizer, String minimizer) {
-			this.maximizer = maximizer;
-			this.minimizer = minimizer;
-		}
-
-		public int score(GameState game, String player) {
-			if (game.hasWon(this.maximizer)) {
-				return +1;
-			} else if (game.hasWon(this.minimizer)) {
-				return -1;
-			} else if (game.isOver()) {
-				return 0;
-			}
-
-			if (this.minimizer.equals(player)) {
-				int minScore = +999;
-				for (String nextMove : game.nextMoves()) {
-					GameState nextGame = game.nextState(nextMove);
-					int score = score(nextGame, this.maximizer);
-					if (score < minScore) {
-						minScore = score;
-					}
-				}
-
-				return minScore;
-			} else {
-				int maxScore = -999;
-				for (String nextMove : game.nextMoves()) {
-					GameState nextGame = game.nextState(nextMove);
-					int score = score(nextGame, this.minimizer);
-					if (score > maxScore) {
-						maxScore = score;
-					}
-				}
-
-				return maxScore;
-			}
-		}
-	}
-
-	public static class GameWithKnownState implements GameState {
+	private static class GameWithKnownState implements GameState {
 		private final boolean isOver;
 		private final String winner;
 		private final Map<String, GameState> nextMoves;
@@ -167,12 +123,5 @@ public class MinimaxSpecs implements SpecClass {
 		public GameState nextState(String move) {
 			return this.nextMoves.get(move);
 		}
-	}
-
-	public interface GameState {
-		boolean hasWon(String player);
-		boolean isOver();
-		List<String> nextMoves();
-		GameState nextState(String move);
 	}
 }
